+++ CHANGELOG.md
@@ -43,6 +43,9 @@ This component adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.h
 
 ### Fixed
 
+- Deadlock: AssemblyResolve_ManagedProfilerDependencies should be moved out
+of Loader class.  [#4269](https://github.com/open-telemetry/opentelemetry-dotnet-instrumentation/issues/4269)
+
 ## [1.12.0](https://github.com/open-telemetry/opentelemetry-dotnet-instrumentation/releases/tag/v1.12.0)
 
 ### Added